+++ packages/deepsea-tools/src/utils/getPostitiveIntParser.ts
@@ -1,0 +1,7 @@
+import { postitiveIntParser } from "./postitiveIntParser"
+
+export function getPostitiveIntParser(defaultValue: number = 1) {
+    return function _postitiveIntParser(value: string | null | undefined) {
+        return postitiveIntParser(value) ?? defaultValue
+    }
+}