+++ packages/deepsea-tools/src/utils/showTotal.ts
@@ -1,0 +1,3 @@
+export function showTotal(total: number) {
+    return `共计 ${total} 条数据`
+}